+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [6.2.0-pre.3] - 2025-04-30
+
+No changes
+
 ## [6.2.0-pre.2] - 2025-04-22
 
 ### Added
@@ -24,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added support for depth preprocessing, when depth data is provided in a linear format.
 - Added occlusion sources flags mask and hands mesh source of occlusion.
+- Added `SerializableGuid.TryGetByteNativeArray` to convert a `SerializableGuid` to a `NativeArray<byte>`.
 
 ### Changed