+++ Documentation~/whats-new.md
@@ -8,5 +8,6 @@ This release includes the following significant changes:
 ## Other API additions
 
 - Added constructors and `defaultValue` properties to `XRSaveAnchorResult`, `XRLoadAnchorResult`, and `XREraseAnchorResult` structs.
+- Added `SerializableGuid.TryGetByteNativeArray` to convert a `SerializableGuid` to a `NativeArray<byte>`.
 
 For a full list of changes in this version including backwards-compatible bugfixes, refer to the package [changelog](xref:arfoundation-changelog).